[#master] refactoring c/grains to get rid of cast to uint64_t

+++ c/grains/src/grains.c
@@ -2,7 +2,7 @@
 
 uint64_t square(uint8_t index)
 {
-	return 0 < index && index <= 64 ? (uint64_t) 1 << (index - 1) : 0;
+	return 0 < index && index <= 64 ? 1UL << (index - 1) : 0;
 }
 
 uint64_t total(void)
